+++ android/app/src/main/res/layout/request_list.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".WalletsFragment">
+    tools:context=".RequestsFragment">
     <data>
         <variable
             name="model"
