+++ android/app/src/main/res/layout/transactions.xml
@@ -3,7 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".WalletsFragment"
+    tools:context=".TransactionsFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
